+++ config/routes.rb
@@ -68,5 +68,7 @@ Rails.application.routes.draw do
       get "payroll", on: :collection
       resources :checks, only: [:create], controller: "claim_checks"
     end
+
+    resources :payroll_runs, only: [:index, :new, :create, :show]
   end
 end